requirements.txt
py==1.4.31
pyTelegramBotAPI==1.4.2
pytest==2.9.0
python-firebase==1.2
requests==2.9.1
six==1.10.0
